+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 
 -   Cleaned up some warnings shown in the console
+-   Fixes an issue where pre api hooks would not work correctly when calling recipe functions directly
 
 ## [0.27.0] - 2022-10-18